Add refactoring and functionality to add empty textfield to category
