+++ front/src/components/Categories/Category.js
@@ -1,8 +1,11 @@
 import React from 'react'
-import { useSelector } from 'react-redux'
+import { useSelector, useDispatch } from 'react-redux'
 
 //Components
 import CategoryItem from './CategoryItem'
+
+//Reducers
+import { setAssignedWord } from '../../reducers/assignReducer'
 
 //Mui components
 import Box from '@mui/material/Box'
@@ -10,6 +13,8 @@ import Grid from '@mui/material/Grid'
 import Typography from '@mui/material/Typography'
 import Divider from '@mui/material/Divider'
 import List from '@mui/material/List'
+import IconButton from '@mui/material/IconButton'
+import AddBoxIcon from '@mui/icons-material/AddBox'
 
 /**
  * Component for category.
@@ -18,6 +23,13 @@ import List from '@mui/material/List'
  */
 const Category = ({ category }) => {
   const assignedWords = useSelector(state => state.assignedWords.filter(word => word.category === category))
+  const dispatch = useDispatch()
+
+  const handleAddNewTextField = () => {
+    console.log('pingpong')
+    console.log(category)
+    dispatch(setAssignedWord('', category))
+  }
 
   return(
     <Grid item>
@@ -30,9 +42,24 @@ const Category = ({ category }) => {
           flexDirection: 'column'
         }}
       >
-        <Typography variant='h6' textAlign='center'>
-          { category }
-        </Typography>
+        <Grid
+          container
+          alignItems='center'
+        >
+          <Grid item xs={8}>
+            <Typography
+              variant='h6'
+              textAlign='center'
+            >
+              {category}
+            </Typography>
+          </Grid>
+          <Grid item xs={4} align='right'>
+            <IconButton onClick={handleAddNewTextField}>
+              <AddBoxIcon />
+            </IconButton>
+          </Grid>
+        </Grid>
         <Divider />
         <Box
           sx={{
